Move implementation to own file
